Move CLI tests out of v1 package
