+++ tests/cli/test_cli_assets.py
