+++ tests/cli/test_cli_auth_guidance.py
