+++ tests/cli/test_cli_open.py
