+++ tests/cli/test_cli_resource_refs.py
