+++ tests/cli/test_cli_simulation_summary.py
