+++ tests/cli/test_remote_resource_logs.py
